test: move conftest
